Reorderd dir structure
