+++ app/controllers/admin/gallery_add.py
@@ -1,14 +1,14 @@
-from app.allImports import *
+from . import admin
 from app.logic.validation import *
-from werkzeug.security import check_password_hash
 from flask import session
 from datetime import datetime
+from flask import render_template
 
-@app.route('/gallery/add', methods=["GET","POST"])
+@admin.route('/gallery/add', methods=["GET","POST"])
 #@require_role('admin')
 def gallery_add():
     if request.method == "GET":
-        return render_template('views/gallery_edit.html')
+        return render_template('views/admin/gallery_edit.html')
     if request.method == "POST":
         galleryData = request.form
         #TODO: need to sanitize description
@@ -20,7 +20,7 @@ def gallery_add():
                                     galleryData['banner'])
         if gid is not None:
             print("create success") #TODO: replace with some flash
-            return redirect(url_for('gallery_edit', gid=gid))
+            return redirect(url_for('administrator.gallery_edit', gid=gid))
         else:
             print("create failed") #TODO: replace with some flash
-            return redirect(url_for('gallery_add'))
+            return redirect(url_for('administrator.gallery_add'))
