+++ app/controllers/public/application_review.py
@@ -4,7 +4,7 @@ from flask import render_template
 
 @public.route('/application/review', methods=["GET","POST"])
 def review():
-    return render_template('views/application_review.html')
+    return render_template('views/public/application_review.html')
