+++ app/templates/views/admin/all_galleries.html
