+++ app/templates/views/admin/gallery_submission.html
