+++ app/templates/views/admin/gallery_view.html
